fix(daemon): unblock gqlgen by reorganizing daemon_ux/worktree_tmux helpers
`make generate` failed on origin/main: gqlgen v0.17.45's "follow-schema" layout
detects existing resolver impls only when they live in `schema.resolvers.go`.
Methods declared in sibling files (`daemon_ux.resolvers.go`,
`worktree_tmux.go`) on `queryResolver` / `subscriptionResolver` /
`worktreeResolver` were re-stubbed into `schema.resolvers.go`, causing
duplicate method declarations on regen.

Compounding: gqlgen mangled multi-paragraph `//`-style doc comments above
those methods — stripping `//` from continuation lines and dumping raw
prose as Go source, breaking gofmt and corrupting the file.

Fix:
- Collapse multi-paragraph `//` doc comments above resolver method
  receivers to single-line form (gqlgen's regen path is reliable here).
- Move the 7 daemon_ux resolver methods (SchemaSdl, WorkView,
  DaemonState, TmuxSessionsChanged, PullRequestChanged, RunChanged,
  WorktreeChanged) and the 2 worktree_tmux resolver methods (TmuxPanes,
  TmuxSession) into `schema.resolvers.go`. The sibling files now hold
  only helper constants/functions.
- Drop the gqlgen-emitted `!!! WARNING !!!` orphan-helper preamble above
  `streamLocalEvents` since the helper is intentional.

Result: `make generate` runs clean and is idempotent. `go build ./...`
succeeds. No new test regressions; pre-existing test failures unchanged.

Closes #526.
